requirements.txt
amqp==2.1.4
asgi-redis==1.0.0
asgiref==1.0.0
attrs==16.3.0
autobahn==17.5.1
Automat==0.5.0
billiard==3.5.0.2
celery==4.0.2
channels==0.17.3
constantly==15.1.0
cx-Oracle==5.3
daphne==0.15.0
Django==1.9.1
django-cors-headers==2.0.2
django-filter==1.0.2
django-grappelli==2.8.3
djangorestframework==3.6.2
docopt==0.6.2
hdfs==2.0.16
incremental==16.10.1
kombu==4.0.2
Markdown==2.6.8
msgpack-python==0.4.8
PyMySQL==0.7.11
pytz==2017.2
redis==2.10.5
requests==2.13.0
six==1.10.0
Twisted==17.1.0
txaio==2.7.1
vine==1.1.3
zope.interface==4.4.0
